+++ services/sso/src/routes/dashboard/usersRoutes.ts
@@ -1,11 +1,12 @@
 import Router from '@koa/router'
-import { authenticateCookie } from '../../middleware/authenticate'
+import { authenticate } from '../../middleware/authenticate'
 import { authorize } from '../../middleware/authorize'
 import { pathRoot } from '../routes'
 import { dashboardListUsers } from '../../controllers/dashboard/users/list'
 import { parse } from '../../middleware/validate'
 import { z } from '../../openapi/zod'
 import { dashboardUsersListQuerySchema } from '../../schemas'
+import { getMe } from '../../controllers/dashboard/users/getMe'
 
 export const dashboardUsersRoutes = new Router()
 
@@ -13,7 +14,7 @@ dashboardUsersRoutes.prefix(pathRoot.v1.dashboard.users)
 
 dashboardUsersRoutes.get(
   '/',
-  authenticateCookie,
+  authenticate,
   authorize('ADMIN'),
   parse(z.object({ query: dashboardUsersListQuerySchema }), {
     useQsParser: true,
@@ -21,3 +22,4 @@ dashboardUsersRoutes.get(
   }),
   dashboardListUsers
 )
+dashboardUsersRoutes.get('/me', authenticate, authorize('ADMIN'), getMe)