+++ web/usuaris/src/constants/urls.ts
@@ -1,4 +1,6 @@
 export const urls = {
-  getUsers: '/api/v1/dashboard/users',
   getItineraries: '/api/v1/itineraries',
+  getUsers: '/api/v1/dashboard/users',
+  logIn: '/api/v1/dashboard/auth/login',
+  getMe: '/api/v1/dashboard/users/me',
 }