+++ web/usuaris/src/hooks/index.ts
@@ -1,2 +1,3 @@
-export { useGetUsers } from './useGetUsers'
 export { useGetItineraries } from './useGetItineraries'
+export { useGetUsers } from './useGetUsers'
+export { useLogin } from './useLogin'